backend/requirements.txt
fastapi
uvicorn
celery
redis
psycopg2-binary
pgvector
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
sentence-transformers
python-multipart
pypdf
beautifulsoup4
requests
python-dotenv